requirements.txt
django==4.1.2
psycopg2-binary==2.9.4
djangorestframework==3.13.1
django-environ==0.9.0
Pillow==9.2.0
django_extensions==3.2.1
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
redis==4.3.4
celery==5.2.7
